database/requirements.txt
psutil==6.0.0
python-dotenv==1.0.1
pandas==2.2.2
numpy==2.0.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
pgvector==0.3.2
pymilvus==2.4.5
sentence-transformers==3.0.1
sentencepiece==0.2.0